change: move schema and queries to cmd/app
this change was made to avoid duplicating the schema.sql file, go:embed
can't access files that are in another right so it had two schema.sql,
one in the root of the project used by sqlc and the other next to
main.go.
has now moved the .sql files to cmd/app and configured the sqlc.yaml
file to use the .sql files that are next to main.